blockifier_reexecution: move compare_block_hash to utils
